Big Update: Api Loading 99%(No test) + refactor logic DB + and more

+++ database/migrations/2024_01_17_074714_create_company_purchases_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->integer('quantity');
-            $table->integer('price');
-            $table->integer('summary');
+            $table->float('price');
+            $table->float('summary');
             $table->foreignId('warehouse_id')->constrained('company_warehouses')->cascadeOnUpdate()->cascadeOnDelete();
             $table->foreignId('plan_account_id')->constrained('plan_accounts')->cascadeOnUpdate();
             $table->foreignId('company_id')->constrained('companies')->cascadeOnUpdate()->cascadeOnDelete();
